requirements.txt
python-telegram-bot>=20.0,<21.0
groq>=0.13.0
requests>=2.31.0
